+++ src/CleanRegex/Match/Details/Detail.php
@@ -44,7 +44,7 @@ interface Detail extends Structure, Intable
 
     public function tail(): int;
 
-    public function textLength(): int;
+    public function length(): int;
 
     public function byteOffset(): int;