+++ src/CleanRegex/Replace/Details/ReplaceDetail.php
@@ -90,9 +90,9 @@ class ReplaceDetail implements Detail
         return $this->detail->text();
     }
 
-    public function textLength(): int
+    public function length(): int
     {
-        return $this->detail->textLength();
+        return $this->detail->length();
     }
 
     public function textByteLength(): int