+++ test/Fakes/CleanRegex/Match/Details/ByteOffsetDetail.php
@@ -29,7 +29,7 @@ class ByteOffsetDetail implements Detail
         throw $this->fail();
     }
 
-    public function textLength(): int
+    public function length(): int
     {
         throw $this->fail();
     }