+++ test/Fakes/CleanRegex/Match/Details/TextDetail.php
@@ -44,7 +44,7 @@ class TextDetail implements Detail
         throw $this->fail();
     }
 
-    public function textLength(): int
+    public function length(): int
     {
         throw $this->fail();
     }