+++ ChangeLog.md
@@ -28,6 +28,7 @@ Incoming in 0.25.0
     * Removed `IntStream.only()`. Use `IntStream.limit().all()` instead.
     * Renamed `Group.textLength()` to `Group.length()`.
     * Renamed `Group.textByteLength()` to `Group.byteLength()`.
+    * Renamed `Detail.textLength()` to `Detail.length()`.
 * Bug fixes
     * Fixed a bug when using `Stream.values().keys().first()` didn't always reindex to `0`.
     * Fixed a bug when using regular groups in `stream()->asInt()` was allowed, but `usingDuplicateName()` groups